Fixed missing profiler color tags due to static initialization order errors
Former-commit-id: 4daa30a9e43a1b96811efee3c62b9a02d8dad2dc [formerly e9ca25b73601381134d441aadb2648665ff4c9a7]
Former-commit-id: e37dfb02f96db85f42386f41621a59fa73e20d31

+++ fancy_core/Annotations.h
@@ -13,14 +13,14 @@ namespace Fancy
 //---------------------------------------------------------------------------//
   namespace Annotations
   {
-    uint8 CreateTag(const char* aName, uint aColor);
-    const AnnotationTagData& GetTagData(uint8 aTag);
+    uint16 CreateTag(const char* aName, uint aColor);
+    const AnnotationTagData& GetTagData(uint16 aTag);
   }
 //---------------------------------------------------------------------------//
 }
 
 #define ANNOTATION_CREATE_TAG(anId, aName, aColor) \
-  const uint8 anId = Fancy::Annotations::CreateTag(aName, aColor);
+  const uint16 anId = Fancy::Annotations::CreateTag(aName, aColor);
 
 #define ANNOTATION_USE_TAG(anId) \
-  extern const uint8 anId;
+  extern const uint16 anId;
